requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.4
Deprecated==1.2.14
distro==1.9.0
elastic-transport==8.12.0
elasticsearch==8.12.1
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.26
llmonitor==0.0.51
load-dotenv==0.1.0
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
pluggy==1.4.0
preshed==3.0.9
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
